refactor: format variables to original state, replace and import into identify_cdg.js
