+++ constants/comme-des-garcons/codeChange.js
@@ -1,2 +1,4 @@
 // pair for the moment that the change of product code structure occured
-export const code_change = ["", 2001];
+const code_change = ['', 2001];
+
+export default code_change;
