+++ constants/comme-des-garcons/lists/exceptionFirstDigitList.js
@@ -1,3 +1,5 @@
 // list of line IDs for which the first digit of the month on the product code is different for lines until 2001
 // take note that there is no overlap between lines that are exceptions and those that are not
-export const exception_firstdigit_list = ["A", "D", "K", "Z"];
+const exception_firstdigit_list = ['A', 'D', 'K', 'Z'];
+
+export default exception_firstdigit_list;
