+++ constants/comme-des-garcons/lists/exceptionLastDigitList.js
@@ -1,3 +1,5 @@
 // list of line IDs for which the first digit of the month on the product code is the last digit of the production year
 // for all lines until 2001
-export const exception_lastdigit_list = ["D", "Z"];
+const exception_lastdigit_list = ['D', 'Z'];
+
+export default exception_lastdigit_list;
