+++ constants/comme-des-garcons/lists/repList.js
@@ -1,5 +1,4 @@
 // list of product code and year pairs that have been spotted on reproductions
-export const rep_list = [
-  ["AZT026", 2005],
-  ["AZT095", 2009],
-];
+const rep_list = [["AZT026", 2005], ["AZT095", 2009]];
+
+export default rep_list;
