+++ constants/comme-des-garcons/lists/sizeList.js
@@ -1,2 +1,4 @@
 // possible characters appearing at the end of the product code used until 2001 (aka monthly code)
-export const size_list = ["", "0", "S", "M", "L"];
+const size_list = ['', '0', 'S', 'M', 'L'];
+
+export default size_list;
