+++ constants/comme-des-garcons/wrongCode.js
@@ -1,2 +1,4 @@
 // code on the care label that is often mistaken for the product code
-export const wrongcode = "DTK9210";
+const wrongcode = "DTK9210";
+
+export default wrongcode;
