+++ constants/yearCurr.js
@@ -1,1 +1,4 @@
-export const year_curr = 2023;
+// current year to act as a bound
+const year_curr = 2023;
+
+export default year_curr;
