+++ dictionary/comme-des-garcons/garmentIdDict.js
@@ -2,19 +2,19 @@
 // categorized alphabetically
 // key: garment ID / value: string for the type of garment
 const garmentID_dict = {
-  A: "a piece",
-  B: "a button-up",
-  C: "a coat",
-  J: "a jacket",
-  K: "a miscellaneous piece",
-  N: "a knitwear piece",
-  O: "a one-piece",
-  P: "pants or shorts",
-  S: "a skirt or a piece from a suit",
-  T: "a top (or track pants)",
-  V: "a vest",
-  X: "a piece",
-  Z: "a piece",
+	'A': "a piece",
+	'B': "a button-up",
+	'C': "a coat",
+	'J': "a jacket",
+	'K': "a miscellaneous piece",
+	'N': "a knitwear piece",
+	'O': "a one-piece",
+	'P': "pants or shorts",
+	'S': "a skirt or a piece from a suit",
+	'T': "a top (or track pants)",
+	'V': "a vest",
+	'X': "a piece",
+	'Z': "a piece"
 };
 
 export default garmentID_dict;
