+++ dictionary/comme-des-garcons/nameDict.js
@@ -2,113 +2,112 @@
 // categorized in (mostly) the same order as the line ID dictionary
 // key: unique short string / value: corresponding name
 const name_dict = {
+
   // _
-  cdg: "COMME des GARÇONS",
+     'cdg':       "COMME des GARÇONS",
   // 1
-  blk: "BLACK COMME des GARÇONS",
+     'blk':       "BLACK COMME des GARÇONS",
   // 2
-  pocket: "COMME des GARÇONS POCKET",
+      'pocket':     "COMME des GARÇONS POCKET",
   // 3
-  taocdg: "tao COMME des GARÇONS",
-  noirkn: "noir kei ninomiya",
+      'taocdg':     "tao COMME des GARÇONS",
+      'noirkn':     "noir kei ninomiya",
   // 5
-  speed: "speedo COMME des GARÇONS",
+      'speed':      "speedo COMME des GARÇONS",
   // 6
-  editw: "COMME des GARÇONS EDITED (for women)",
+      'editw':      "COMME des GARÇONS EDITED (for women)",
   // 7
-  editm: "COMME des GARÇONS EDITED (for men)",
+      'editm':      "COMME des GARÇONS EDITED (for men)",
   // 8
-  wal: "Wallet COMME des GARÇONS",
+      'wal':        "Wallet COMME des GARÇONS",
   // A
-  noir: "COMME des GARÇONS NOIR",
-  play: "PLAY COMME des GARÇONS",
+      'noir':       "COMME des GARÇONS NOIR",
+      'play':       "PLAY COMME des GARÇONS",
   // B
-  ppDSM:
-    "COMME des GARÇONS parfums PARFUMS (shop exclusive DOVER STREET MARKET launch)",
+      'ppDSM':      "COMME des GARÇONS parfums PARFUMS (shop exclusive DOVER STREET MARKET launch)",
   // D
-  h2: "COMME des GARÇONS HOMME DEUX",
+      'h2':         "COMME des GARÇONS HOMME DEUX",
   // E
-  rkcdg: "Rei Kawakubo (COMME des GARÇONS canadian trademark)",
-  gan: "GANRYU COMME des GARÇONS",
+      'rkcdg':      "Rei Kawakubo (COMME des GARÇONS canadian trademark)",
+      'gan':        "GANRYU COMME des GARÇONS",
   // F
-  shirt: "COMME des GARÇONS SHIRT",
-  shirtgirl: "COMME des GARÇONS SHIRT (GIRL)",
-  shirtboy: "COMME des GARÇONS SHIRT BOY",
-  shirtboys: "COMME des GARÇONS SHIRT boys",
-  shirtFO: "COMME des GARÇONS SHIRT (FOREVER)",
+      'shirt':      "COMME des GARÇONS SHIRT",
+      'shirtgirl':  "COMME des GARÇONS SHIRT (GIRL)",
+      'shirtboy':   "COMME des GARÇONS SHIRT BOY",
+      'shirtboys':  "COMME des GARÇONS SHIRT boys",
+      'shirtFO':     "COMME des GARÇONS SHIRT (FOREVER)",
   // H
-  hcdg: "HOMME COMME des GARÇONS",
-  h: "COMME des GARÇONS HOMME",
+      'hcdg':       "HOMME COMME des GARÇONS",
+      'h':          "COMME des GARÇONS HOMME",
   // I
-  ind: "COMME des GARÇONS indigo",
-  hi: "COMME des GARÇONS HOMME (inverted)",
-  hh: "COMME des GARÇONS HOMME HOMME",
-  cdgSE: "COMME des GARÇONS (shop exclusive)",
-  alive: "MAGAZINE ALIVE (VOGUE NIPPON and COMME des GARÇONS)",
-  gds: "GOOD DESIGN SHOP COMME des GARÇONS",
+      'ind':        "COMME des GARÇONS indigo",
+      'hi':         "COMME des GARÇONS HOMME (inverted)",
+      'hh':         "COMME des GARÇONS HOMME HOMME",
+      'cdgSE':      "COMME des GARÇONS (shop exclusive)",
+      'alive':      "MAGAZINE ALIVE (VOGUE NIPPON and COMME des GARÇONS)",
+      'gds':        "GOOD DESIGN SHOP COMME des GARÇONS",
   // J
-  jw: "JUNYA WATANABE COMME des GARÇONS",
+      'jw':         "JUNYA WATANABE COMME des GARÇONS",
   // K
-  aoy: "COMME des GARÇONS (Aoyama exclusive)",
-  chaoy: "COMME des GARÇONS Chrome Hearts (Aoyama exclusive)",
+      'aoy':        "COMME des GARÇONS (Aoyama exclusive)",
+      'chaoy':      "COMME des GARÇONS Chrome Hearts (Aoyama exclusive)",
   // L
-  ise115: "COMME des GARÇONS (ISETAN 115yrs)",
-  ise2man:
-    "COMME des GARÇONS HOMME PLUS (CELEBRATION ISETAN MEN'S 1YEAR + COMME des GARCONS 2 MAN)",
-  ppoh: "COMME des GARÇONS parfums PARFUMS (shop exclusive ODOUR HOLIDAY)",
+      'ise115':     "COMME des GARÇONS (ISETAN 115yrs)",
+      'ise2man':    "COMME des GARÇONS HOMME PLUS (CELEBRATION ISETAN MEN'S 1YEAR + COMME des GARCONS 2 MAN)",
+      'ppoh':       "COMME des GARÇONS parfums PARFUMS (shop exclusive ODOUR HOLIDAY)",
   // N
-  girl: "COMME des GARÇONS GIRL",
+      'girl':       "COMME des GARÇONS GIRL",
   // O
-  "10ccno": "10 corso como COMME des GARÇONS (shop exclusive number series)",
-  aoySE: "COMME des GARÇONS (Aoyama exclusive)",
-  walSE: "Wallet COMME des GARÇONS (shop exclusive)",
-  eyeSE: "eYe JUNYA WATANABE COMME des GARÇONS MAN (shop exclusive)",
-  speedSE: "speedo COMME des GARÇONS (shop exclusive)",
-  trad: "TRADING MUSEUM COMME des GARÇONS",
-  playxmas: "PLAY COMME des GARÇONS (shop exclusive holidays special)",
-  bm: "black market COMME des GARÇONS",
-  CDGSE: "CDG (shop exclusive)",
-  friends: "Friends and COMME des GARÇONS Happy Holidays",
-  people:
-    "COMME des GARÇONS (People of the Year by COMME des GARÇONS and SWITCH)",
+      '10ccno':     "10 corso como COMME des GARÇONS (shop exclusive number series)",
+      'aoySE':      "COMME des GARÇONS (Aoyama exclusive)",
+      'walSE':      "Wallet COMME des GARÇONS (shop exclusive)",
+      'eyeSE':      "eYe JUNYA WATANABE COMME des GARÇONS MAN (shop exclusive)",
+      'speedSE':    "speedo COMME des GARÇONS (shop exclusive)",
+      'trad':       "TRADING MUSEUM COMME des GARÇONS",
+      'playxmas':   "PLAY COMME des GARÇONS (shop exclusive holidays special)",
+      'bm':         "black market COMME des GARÇONS",
+      'CDGSE':      "CDG (shop exclusive)",
+      'friends':    "Friends and COMME des GARÇONS Happy Holidays",
+      'people':     "COMME des GARÇONS (People of the Year by COMME des GARÇONS and SWITCH)",
   // P
-  hp: "COMME des GARÇONS HOMME PLUS",
-  hps: "SPORT COMME des GARÇONS HOMME PLUS",
-  hpeg: "COMME des GARÇONS HOMME PLUS EVER GREEN",
+      'hp':         "COMME des GARÇONS HOMME PLUS",
+      'hps':        "SPORT COMME des GARÇONS HOMME PLUS",
+      'hpeg':       "COMME des GARÇONS HOMME PLUS EVER GREEN",
   // R
-  rdc: "robe de chambre COMME des GARÇONS",
-  cdgcdg: "COMME des GARÇONS COMME des GARÇONS",
+      'rdc':        "robe de chambre COMME des GARÇONS",
+      'cdgcdg':     "COMME des GARÇONS COMME des GARÇONS",
   // S
-  "10cc": "10 corso como COMME des GARÇONS",
-  CDG: "CDG",
+      '10cc':       "10 corso como COMME des GARÇONS",
+      'CDG':       "CDG",
   // T
-  tric: "tricot COMME des GARÇONS",
-  tao: "tao",
+      'tric':       "tricot COMME des GARÇONS",
+      'tao':        "tao",
   // U
-  hpd: "COMME des GARÇONS HOMME PLUS in collaboration with District UNITED ARROWS",
-  pink: "JUNYA WATANABE COMME des GARÇONS (pink)",
+      'hpd':        "COMME des GARÇONS HOMME PLUS in collaboration with District UNITED ARROWS",
+      'pink':       "JUNYA WATANABE COMME des GARÇONS (pink)",
   // V
-  beat: "The Beatles COMME des GARÇONS",
+      'beat':       "The Beatles COMME des GARÇONS",
   // W
-  jwman: "JUNYA WATANABE COMME des GARÇONS MAN",
-  eye: "eYe JUNYA WATANABE COMME des GARÇONS MAN",
+      'jwman':      "JUNYA WATANABE COMME des GARÇONS MAN",
+      'eye':        "eYe JUNYA WATANABE COMME des GARÇONS MAN",
   // X
-  rkhp: "Rei Kawakubo (COMME des GARÇONS HOMME PLUS canadian trademark)",
-  jwd: "JUNYA WATANABE COMME des GARÇONS denim",
+      'rkhp':       "Rei Kawakubo (COMME des GARÇONS HOMME PLUS canadian trademark)",
+      'jwd':        "JUNYA WATANABE COMME des GARÇONS denim",
   // Z
-  two: "COMME des GARÇONS HOMME HOMME (Aoyama exclusive COMME des GARÇONS two)",
-  peg: "Peggy Moffitt COMME des GARÇONS",
-  cdgch: "COMME des GARÇONS Chrome Hearts",
+      'two':        "COMME des GARÇONS HOMME HOMME (Aoyama exclusive COMME des GARÇONS two)",
+      'peg':        "Peggy Moffitt COMME des GARÇONS",
+      'cdgch':      "COMME des GARÇONS Chrome Hearts",
+
 
   // EXCEPTIONS
-  "10ccBB": "10 corso como COMME des GARÇONS (BE@RBRICK series)",
-  hpch: "COMME des GARÇONS HOMME PLUS customized by Chrome Hearts", // both labels on garment
-  "10ccSS": "10 corso como COMME des GARÇONS (summer series)",
-  shchre: "COMME des GARÇONS Chrome Hearts (striped shirt limited edition)", // only the double name label
-  shch: "COMME des GARÇONS SHIRT customized by Chrome Hearts", // both labels on garment
-  hpchre:
-    "COMME des GARÇONS Chrome Hearts (custom COMME des GARÇONS HOMME PLUS)", // only the double name label
-  chlim: "COMME des GARÇONS Chrome Hearts (Aoyama exclusive limited release)", // only the double name label
+      '10ccBB':     "10 corso como COMME des GARÇONS (BE@RBRICK series)",
+      'hpch':       "COMME des GARÇONS HOMME PLUS customized by Chrome Hearts", // both labels on garment
+      '10ccSS':     "10 corso como COMME des GARÇONS (summer series)",
+      'shchre':     "COMME des GARÇONS Chrome Hearts (striped shirt limited edition)", // only the double name label
+      'shch':       "COMME des GARÇONS SHIRT customized by Chrome Hearts",  // both labels on garment
+      'hpchre':     "COMME des GARÇONS Chrome Hearts (custom COMME des GARÇONS HOMME PLUS)", // only the double name label
+      'chlim':      "COMME des GARÇONS Chrome Hearts (Aoyama exclusive limited release)", // only the double name label
+
 };
 
 export default name_dict;
