+++ identify_cdg.js
@@ -1,3 +1,9 @@
+import year_curr from "./constants/yearCurr";
+import { code_change, wrongcode } from "./constants/comme-des-garcons"
+import { exception_firstdigit_list, exception_lastdigit_list, rep_list, size_list } from "./constants/comme-des-garcons/lists"
+import { exception_code_dict, garmentID_dict, lineID_dict, name_dict, seasonID_dict } from "./dictionary/comme-des-garcons"
+
+
 ////////////////////////////////////////////////////////////////////
 ///// WELCOME TO THE 'COMME des GARÇONS' IDENTIFICATION CENTER /////
 ////////////////////////////////////////////////////////////////////
@@ -8,350 +14,6 @@
 
 // This program was conceived by the genius behind My Clothing Archive
 // and it was built according to the COMME des GARÇONS identification chart.
-
-
-
-///// DICTIONARIES & OTHER VARIABLES
-
-// current year to act as a bound
-// const year_curr = 2023;
-
-// name dictionary for the different lines (and details) that are attached to different line IDs
-// categorized in (mostly) the same order as the line ID dictionary
-// key: unique short string / value: corresponding name
-const name_dict = {
-
-    // _
-       'cdg':       "COMME des GARÇONS",
-    // 1
-       'blk':       "BLACK COMME des GARÇONS",
-    // 2
-        'pocket':     "COMME des GARÇONS POCKET",
-    // 3
-        'taocdg':     "tao COMME des GARÇONS",
-        'noirkn':     "noir kei ninomiya",
-    // 5
-        'speed':      "speedo COMME des GARÇONS",
-    // 6
-        'editw':      "COMME des GARÇONS EDITED (for women)",
-    // 7
-        'editm':      "COMME des GARÇONS EDITED (for men)",
-    // 8
-        'wal':        "Wallet COMME des GARÇONS",
-    // A
-        'noir':       "COMME des GARÇONS NOIR",
-        'play':       "PLAY COMME des GARÇONS",
-    // B
-        'ppDSM':      "COMME des GARÇONS parfums PARFUMS (shop exclusive DOVER STREET MARKET launch)",
-    // D
-        'h2':         "COMME des GARÇONS HOMME DEUX",
-    // E
-        'rkcdg':      "Rei Kawakubo (COMME des GARÇONS canadian trademark)",
-        'gan':        "GANRYU COMME des GARÇONS",
-    // F
-        'shirt':      "COMME des GARÇONS SHIRT",
-        'shirtgirl':  "COMME des GARÇONS SHIRT (GIRL)",
-        'shirtboy':   "COMME des GARÇONS SHIRT BOY",
-        'shirtboys':  "COMME des GARÇONS SHIRT boys",
-        'shirtFO':     "COMME des GARÇONS SHIRT (FOREVER)",
-    // H
-        'hcdg':       "HOMME COMME des GARÇONS",
-        'h':          "COMME des GARÇONS HOMME",
-    // I
-        'ind':        "COMME des GARÇONS indigo",
-        'hi':         "COMME des GARÇONS HOMME (inverted)",
-        'hh':         "COMME des GARÇONS HOMME HOMME",
-        'cdgSE':      "COMME des GARÇONS (shop exclusive)",
-        'alive':      "MAGAZINE ALIVE (VOGUE NIPPON and COMME des GARÇONS)",
-        'gds':        "GOOD DESIGN SHOP COMME des GARÇONS",
-    // J
-        'jw':         "JUNYA WATANABE COMME des GARÇONS",
-    // K
-        'aoy':        "COMME des GARÇONS (Aoyama exclusive)",
-        'chaoy':      "COMME des GARÇONS Chrome Hearts (Aoyama exclusive)",
-    // L
-        'ise115':     "COMME des GARÇONS (ISETAN 115yrs)",
-        'ise2man':    "COMME des GARÇONS HOMME PLUS (CELEBRATION ISETAN MEN'S 1YEAR + COMME des GARCONS 2 MAN)",
-        'ppoh':       "COMME des GARÇONS parfums PARFUMS (shop exclusive ODOUR HOLIDAY)",
-    // N
-        'girl':       "COMME des GARÇONS GIRL",
-    // O
-        '10ccno':     "10 corso como COMME des GARÇONS (shop exclusive number series)",
-        'aoySE':      "COMME des GARÇONS (Aoyama exclusive)",
-        'walSE':      "Wallet COMME des GARÇONS (shop exclusive)",
-        'eyeSE':      "eYe JUNYA WATANABE COMME des GARÇONS MAN (shop exclusive)",
-        'speedSE':    "speedo COMME des GARÇONS (shop exclusive)",
-        'trad':       "TRADING MUSEUM COMME des GARÇONS",
-        'playxmas':   "PLAY COMME des GARÇONS (shop exclusive holidays special)",
-        'bm':         "black market COMME des GARÇONS",
-        'CDGSE':      "CDG (shop exclusive)",
-        'friends':    "Friends and COMME des GARÇONS Happy Holidays",
-        'people':     "COMME des GARÇONS (People of the Year by COMME des GARÇONS and SWITCH)",
-    // P
-        'hp':         "COMME des GARÇONS HOMME PLUS",
-        'hps':        "SPORT COMME des GARÇONS HOMME PLUS",
-        'hpeg':       "COMME des GARÇONS HOMME PLUS EVER GREEN",
-    // R
-        'rdc':        "robe de chambre COMME des GARÇONS",
-        'cdgcdg':     "COMME des GARÇONS COMME des GARÇONS",
-    // S
-        '10cc':       "10 corso como COMME des GARÇONS",
-        'CDG':       "CDG",
-    // T
-        'tric':       "tricot COMME des GARÇONS",
-        'tao':        "tao",
-    // U
-        'hpd':        "COMME des GARÇONS HOMME PLUS in collaboration with District UNITED ARROWS",
-        'pink':       "JUNYA WATANABE COMME des GARÇONS (pink)",
-    // V
-        'beat':       "The Beatles COMME des GARÇONS",
-    // W
-        'jwman':      "JUNYA WATANABE COMME des GARÇONS MAN",
-        'eye':        "eYe JUNYA WATANABE COMME des GARÇONS MAN",
-    // X
-        'rkhp':       "Rei Kawakubo (COMME des GARÇONS HOMME PLUS canadian trademark)",
-        'jwd':        "JUNYA WATANABE COMME des GARÇONS denim",
-    // Z
-        'two':        "COMME des GARÇONS HOMME HOMME (Aoyama exclusive COMME des GARÇONS two)",
-        'peg':        "Peggy Moffitt COMME des GARÇONS",
-        'cdgch':      "COMME des GARÇONS Chrome Hearts",
-
-
-    // EXCEPTIONS
-        '10ccBB':     "10 corso como COMME des GARÇONS (BE@RBRICK series)",
-        'hpch':       "COMME des GARÇONS HOMME PLUS customized by Chrome Hearts", // both labels on garment
-        '10ccSS':     "10 corso como COMME des GARÇONS (summer series)",
-        'shchre':     "COMME des GARÇONS Chrome Hearts (striped shirt limited edition)", // only the double name label
-        'shch':       "COMME des GARÇONS SHIRT customized by Chrome Hearts",  // both labels on garment
-        'hpchre':     "COMME des GARÇONS Chrome Hearts (custom COMME des GARÇONS HOMME PLUS)", // only the double name label
-        'chlim':      "COMME des GARÇONS Chrome Hearts (Aoyama exclusive limited release)", // only the double name label
-
-};
-
-// line ID dictionary for the different line IDs and their information (name, schedule, and span of existence)
-// categorized alphabetically and chronologically
-// key: line ID / value: [list of possible [name, type of schedule, [first [season, year] pair, last [season, year] pair]]]
-    // the type of schedule relates to the production schedule during the year
-        // high-end lines are generally produced early ('E')
-        // auxiliary lines are generally produced late ('L')
-        // meaningless for lines that appeared in 2001 or later
-    // 'S' is for Spring/Summer and 'W' is for Autumn/Winter
-    // if it is unclear or if it is not a seasonal matter, the season is left as an empty string
-    // if the line is still ongoing, the season is left as an empty string and the year is set to 9999
-const lineID_dict = {
-
-    '_': [[name_dict['cdg'],        'L',    [['',  1969], ['S', 1982]]]],
-
-    '1': [[name_dict['blk'],        'L',    [['S', 2009], ['',  9999]]]],
-
-    '2': [[name_dict['pocket'],     'L',    [['S', 2011], ['',  2017]]]],
-
-    '3': [[name_dict['taocdg'],     'E',    [['W', 2005], ['S', 2011]]],
-          [name_dict['noirkn'],     'E',    [['W', 2013], ['',  9999]]]],
-
-    '5': [[name_dict['speed'],      'L',    [['S', 2007], ['S', 2009]]]],
-
-    '6': [[name_dict['editw'],      'L',    [['S', 2007], ['',  9999]]]],
-
-    '7': [[name_dict['editm'],      'L',    [['S', 2007], ['',  9999]]]],
-
-    '8': [[name_dict['wal'],        'L',    [['',  1980], ['',  9999]]]],
-
-    'A': [[name_dict['noir'],       'E',    [['W', 1987], ['W', 1995]]],
-          [name_dict['play'],       'L',    [['',  2002], ['',  9999]]]],
-
-    'B': [[name_dict['ppDSM'],      'L',    [['W', 2009], ['W', 2009]]]],
-
-    'D': [[name_dict['h2'],         'L',    [['',  1987], ['',  9999]]]],
-
-    'E': [[name_dict['cdg'],        'E',    [['W', 1982], ['W', 2005]]],
-          [name_dict['rkcdg'],      'E',    [['S', 1984], ['W', 1995]]],
-          [name_dict['noir'],       'E',    [['W', 1987], ['W', 1995]]],
-          [name_dict['gan'],        'E',    [['S', 2008], ['S', 2017]]]],
-
-    'F': [[name_dict['shirtFO'],    'L',    [['W', 2020], ['',  9999]]],
-          [name_dict['shirt'],      'L',    [['S', 2021], ['',  9999]]]],
-
-    'G': [[name_dict['cdg'],        'E',    [['W', 1982], ['',  9999]]],
-          [name_dict['noir'],       'E',    [['W', 2005], ['S', 2009]]],
-          [name_dict['bm'],         'E',    [['S', 2017], ['',  9999]]]],
-
-    'H': [[name_dict['hcdg'],       'L',    [['',  1978], ['W', 1983]]],
-          [name_dict['h'],          'L',    [['S', 1984], ['',  9999]]]],
-
-    'I': [[name_dict['ind'],        'L',    [['S', 1984], ['W', 1987]]],
-          [name_dict['hi'],         'L',    [['S', 1994], ['W', 1998]]],
-          [name_dict['hh'],         'L',    [['S', 1999], ['W', 2001]]],
-          [name_dict['jwman'],      'L',    [['S', 2006], ['S', 2006]]],
-          [name_dict['alive'],      'L',    [['W', 2009], ['W', 2009]]],
-          [name_dict['cdgSE'],      'L',    [['W', 2011], ['W', 2012]]],
-          [name_dict['gds'],        'L',    [['',  2011], ['',  2018]]]],
-
-    'J': [[name_dict['jw'],         'E',    [['W', 1992], ['',  9999]]]],
-
-    'K': [[name_dict['aoy'],        'E',    [['S', 1989], ['',  9999]]],
-          [name_dict['chaoy'],      'E',    [['S', 2012], ['W', 2012]]]],
-
-    'L': [[name_dict['ise115'],     'L',    [['W', 2001], ['W', 2001]]],
-          [name_dict['ise2man'],    'L',    [['W', 2004], ['W', 2004]]],
-          [name_dict['ppoh'],       'L',    [['W', 2019], ['W', 2019]]]],
-
-    'N': [[name_dict['girl'],       'L',    [['S', 2015], ['',  9999]]]],
-
-    'O': [[name_dict['10ccno'],     'L',    [['S', 2002], ['W', 2003]]],
-          [name_dict['aoySE'],      'E',    [['W', 2004], ['W', 2004]]],
-          [name_dict['cdgcdg'],     'L',    [['S', 2005], ['S', 2005]]],
-          [name_dict['cdgSE'],      'L',    [['',  2005], ['',  9999]]],
-          [name_dict['walSE'],      'L',    [['',  2005], ['',  9999]]],
-          [name_dict['eyeSE'],      'L',    [['W', 2005], ['W', 2005]]],
-          [name_dict['speedSE'],    'L',    [['S', 2006], ['S', 2006]]],
-          [name_dict['trad'],       'L',    [['W', 2009], ['',  9999]]],
-          [name_dict['playxmas'],   'L',    [['W', 2013], ['W', 2017]]],
-          [name_dict['bm'],         'L',    [['S', 2017], ['',  9999]]],
-          [name_dict['CDGSE'],      'L',    [['S', 2018], ['S', 2018]]],
-          [name_dict['friends'],    'L',    [['W', 2018], ['W', 2018]]],
-          [name_dict['ppoh'],       'L',    [['W', 2019], ['W', 2019]]],
-          [name_dict['people'],     'L',    [['W', 2020], ['W', 2020]]]],
-
-    'P': [[name_dict['hp'],         'E',    [['S', 1985], ['',  9999]]],
-          [name_dict['hps'],        'E',    [['W', 2005], ['W', 2006]]],
-          [name_dict['hpeg'],       'E',    [['W', 2005], ['W', 2009]]]],
-
-    'R': [[name_dict['rdc'],        'L',    [['',  1981], ['W', 2004]]],
-          [name_dict['cdgcdg'],     'L',    [['S', 2005], ['',  9999]]]],
-
-    'S': [[name_dict['10cc'],       'L',    [['S', 2002], ['W', 2011]]],
-          [name_dict['CDG'],        'L',    [['S', 2018], ['',  9999]]]],
-
-    'T': [[name_dict['tric'],       'L',    [['',  1981], ['W', 2021]]],
-          [name_dict['tao'],        'E',    [['S', 2022], ['',  9999]]]],
-
-    'U': [[name_dict['hpd'],        'E',    [['W', 2000], ['W', 2005]]],
-          [name_dict['pink'],       'L',    [['S', 2003], ['S', 2014]]]],
-
-    'V': [[name_dict['jw'],         'E',    [['W', 1992], ['W', 2005]]],
-          [name_dict['beat'],       'L',    [['S', 2009], ['',  9999]]]],
-
-    'W': [[name_dict['jwman'],      'E',    [['S', 2002], ['',  9999]]],
-          [name_dict['eye'],        'L',    [['W', 2005], ['',  9999]]]],
-
-    'X': [[name_dict['hp'],         'E',    [['S', 1985], ['W', 2005]]],
-          [name_dict['rkhp'],       'E',    [['S', 1985], ['W', 1995]]],
-          [name_dict['jwd'],        'L',    [['',  2014], ['W', 2021]]]],
-
-    'Y': [[name_dict['jwman'],      'E',    [['S', 2002], ['W', 2005]]]],
-
-    'Z': [[name_dict['two'],        'L',    [['S', 1999], ['W', 2001]]],
-          [name_dict['peg'],        'L',    [['S', 2003], ['W', 2005]]],
-          [name_dict['cdgch'],      'L',    [['',  2012], ['',  2018]]]]
-
-};
-
-// list of product code and year pairs that have been spotted on reproductions
-const rep_list = [["AZT026", 2005], ["AZT095", 2009]];
-
-// code on the care label that is often mistaken for the product code
-const wrongcode = "DTK9210";
-
-// exception dictionary for the different product code exceptions and their information (stylized code, garment ID, name, and point in time)
-// categorized chronologically
-// key: product code exception / value: [list of possible [stylized code, garment ID, [list of possible [name, [list of possible [season, year] pairs]]]]]
-    // 'S' is for Spring/Summer and 'W' is for Autumn/Winter
-    // if it is unclear or if it is not a seasonal matter, the season is left as an empty string
-const exception_code_dict = {
-    "08BBT01":          ["08BBT01",         0,      'T',    '',     [[name_dict['10ccBB'],  [['',  2004]]]]],
-    "08BBT02":          ["08BBT02",         0,      'T',    '',     [[name_dict['10ccBB'],  [['',  2004]]]]],
-    "BEARBRICKT001":    ["BEARBRICK-T001",  2004,   'T',    '',     [[name_dict['10ccBB'],  [['',  2004]]]]],
-    "BEARBRICKT002":    ["BEARBRICK-T002",  2004,   'T',    '',     [[name_dict['10ccBB'],  [['',  2004]]]]],
-    "BEARBRICKT002":    ["BEARBRICK-T003",  2004,   'T',    '',     [[name_dict['10ccBB'],  [['',  2004]]]]],
-    "BEARBRICKT004":    ["BEARBRICK-T004",  2004,   'T',    '',     [[name_dict['10ccBB'],  [['',  2004]]]]],
-    "PSJ063":           ["PS-J063",         2006,   'J',    '',     [[name_dict['hpch'],    [['W', 2007]]]]],
-    "OTJ007":           ["OT-J007",         2007,   'J',    '',     [[name_dict['hpch'],    [['W', 2007]]]]],
-    "SD19T60C":         ["SD19T60C",        0,      'T',    '',     [[name_dict['speed'],   [['S', 2009]]]]],
-    "5FT08A":           ["5FT08A",          0,      'T',    '',     [[name_dict['10ccSS'],  [['S', 2011]]]]],
-    "5FT08B":           ["5FT08B",          0,      'T',    '',     [[name_dict['10ccSS'],  [['S', 2011]]]]],
-    "CDGS2ST":          ["CDGS2ST",         0,      'B',    '',     [[name_dict['shchre'],  [['',  2013]]],
-                                                                     [name_dict['shch'],    [['',  2014],
-                                                                                             ['',  2015]]]]],
-    "CDGS2PL":          ["CDGS2PL",         0,      'B',    '',     [[name_dict['shch'],    [['',  2015]]]]],
-    "PDB007":           ["PD-B007",         2019,   'B',    '',     [[name_dict['hpch'],    [['W', 2021]]]]],
-    "PFP043":           ["PF-P043",         2020,   'P',    '',     [[name_dict['hpch'],    [['W', 2021]]]]],
-    "CDGCH01":          ["CDG-CH-01",       0,      'B',    '',     [[name_dict['chlim'],   [['W', 2021]]]]],
-    "CDGCH02":          ["CDG-CH-02",       0,      'B',    '',     [[name_dict['chlim'],   [['W', 2021]]]]],
-    "CDGCH03":          ["CDG-CH-03",       0,      'B',    '',     [[name_dict['chlim'],   [['W', 2021]]]]],
-    "CDGCH04":          ["CDG-CH-04",       0,      'B',    '',     [[name_dict['chlim'],   [['W', 2021]]]]],
-};
-
-// list of line IDs for which the first digit of the month on the product code is different for lines until 2001
-// take note that there is no overlap between lines that are exceptions and those that are not
-const exception_firstdigit_list = ['A', 'D', 'K', 'Z'];
-
-// list of line IDs for which the first digit of the month on the product code is the last digit of the production year
-// for all lines until 2001
-const exception_lastdigit_list = ['D', 'Z'];
-
-// pair for the moment that the change of product code structure occured
-const code_change = ['', 2001];
-
-// season ID dictionary for the different season IDs and their corresponding season (if one) and year pairs
-// categorized alphabetically
-// key: season ID / value: [list of [season, year] pairs]
-    // 'S' is for Spring/Summer and 'W' is for Autumn/Winter
-    // for U and Z, there is no specific season and year pair they refer to
-const seasonID_dict = {
-    'A':  [             ['S', 2008], ['S', 2018]],
-    'B':  [             ['W', 2008], ['W', 2018]],
-    'C':  [['W', 2001], ['S', 2009], ['S', 2019]],
-    'D':  [['W', 2001], ['W', 2009], ['W', 2019]],
-    'E':  [['S', 2002], ['S', 2010], ['S', 2020]],
-    'F':  [['S', 2002], ['W', 2010], ['W', 2020]],
-    'G':  [['W', 2002], ['S', 2011], ['S', 2021]],
-    'H':  [['W', 2002], ['W', 2011], ['W', 2021]],
-    'I':  [['S', 2003], ['S', 2012], ['S', 2022]],
-    'J':  [['S', 2003], ['W', 2012], ['W', 2022]],
-    'K':  [['W', 2003], ['S', 2013], ['S', 2023]],
-    'L':  [['W', 2003], ['W', 2013]],
-    'M':  [['S', 2004], ['S', 2014]],
-    'N':  [['W', 2004], ['W', 2014]],
-    'O':  [['S', 2005], ['S', 2015]],
-    'P':  [['W', 2005], ['W', 2015]],
-    'Q':  [['S', 2006], ['S', 2016]],
-    'R':  [['W', 2006], ['W', 2016]],
-    'S':  [['S', 2007], ['S', 2017]],
-    'T':  [['W', 2007], ['W', 2017]],
-    'U':  [['',  2001], ['',  2002], ['',  2003], ['',  2004], ['',  2005], ['',  2006], ['',  2007], ['',  2008], ['',  2009], ['',  2010],
-           ['',  2011], ['',  2012], ['',  2013], ['',  2014], ['',  2015], ['',  2016], ['',  2017], ['',  2018], ['',  2019], ['',  2020],
-           ['',  2021], ['',  2022], ['',  2023]],
-    'Z':  [['',  2001], ['',  2002], ['',  2003], ['',  2004], ['',  2005], ['',  2006], ['',  2007], ['',  2008], ['',  2009], ['',  2010],
-           ['',  2011], ['',  2012], ['',  2013], ['',  2014], ['',  2015], ['',  2016], ['',  2017], ['',  2018], ['',  2019], ['',  2020],
-           ['',  2021], ['',  2022], ['',  2023]],
-
-    'LC': [['W', 2019]],
-    'PP': [['W', 2005], ['S', 2006], ['W', 2015]]
-};
-
-// garment ID dictionary for the different garment IDs and their corresponding type of garment
-// categorized alphabetically
-// key: garment ID / value: string for the type of garment
-const garmentID_dict = {
-	'A': "a piece",
-	'B': "a button-up",
-	'C': "a coat",
-	'J': "a jacket",
-	'K': "a miscellaneous piece",
-	'N': "a knitwear piece",
-	'O': "a one-piece",
-	'P': "pants or shorts",
-	'S': "a skirt or a piece from a suit",
-	'T': "a top (or track pants)",
-	'V': "a vest",
-	'X': "a piece",
-	'Z': "a piece"
-};
-
-// possible characters appearing at the end of the product code used until 2001 (aka monthly code)
-const size_list = ['', '0', 'S', 'M', 'L'];
-
 
 ///// ERROR FUNCTIONS
